requirements.txt
python-vlc==3.0.6108
pydub==0.23.1